// src/core/api/generated/spring/schemas/ProductVariantSelectionDTOStatus.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Crm Backend API
 * Crm Backend API documentation
 * OpenAPI spec version: 0.0.1
 */

export type ProductVariantSelectionDTOStatus = typeof ProductVariantSelectionDTOStatus[keyof typeof ProductVariantSelectionDTOStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ProductVariantSelectionDTOStatus = {
  DRAFT: 'DRAFT',
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE',
  ARCHIVED: 'ARCHIVED',
} as const;