// src/core/api/generated/spring/schemas/SystemConfigAttributeDTO.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Crm Backend API
 * Crm Backend API documentation
 * OpenAPI spec version: 0.0.1
 */
import type { SystemConfigAttributeDTOAttributeType } from './SystemConfigAttributeDTOAttributeType';
import type { SystemConfigAttributeDTOStatus } from './SystemConfigAttributeDTOStatus';
import type { SystemConfigDTO } from './SystemConfigDTO';

export interface SystemConfigAttributeDTO {
  id?: number;
  /**
   * @minLength 1
   * @maxLength 50
   * @pattern ^[a-z][a-z0-9_]*$
   */
  name: string;
  /**
   * @minLength 1
   * @maxLength 100
   */
  label: string;
  attributeType: SystemConfigAttributeDTOAttributeType;
  isRequired: boolean;
  /** @minimum 0 */
  sortOrder: number;
  status: SystemConfigAttributeDTOStatus;
  /**
   * @minLength 0
   * @maxLength 50
   */
  createdBy?: string;
  createdDate?: string;
  /**
   * @minLength 0
   * @maxLength 50
   */
  lastModifiedBy?: string;
  lastModifiedDate?: string;
  systemConfig: SystemConfigDTO;
}